Reduce shimming of GpuFlatMapGroupsInPandasExec
